requirements.txt
dj-database-url==0.4.2
Django==1.11.4
djangorestframework==3.6.4
psycopg2==2.7.3.1
pytz==2017.2
whitenoise==3.3.0
